requirements.txt
pysftp
Django==1.6.1
django-jenkins
south
#vectorformats
simplejson
openpyxl==1.5.8
xlrd
pysolr
pep8
mox
mock
python-dateutil
django_coverage
django-userena==1.2.1
django-guardian==1.1.1
django-nose
#python-memcached
#django-memcached
django-tastypie==0.9.15
httplib2
model_mommy==0.8.1
django-waffle
nose-exclude
django-dajaxice
django-dajax
django-bootstrap-toolkit
requests==2.4.3
#numpy==1.8.0
gunicorn
thumbor
pillow